cmd/src: refactor command logic to support nested subcommands
